refactor: extract context module glob handling
